fix(instrumentation): Align release deployment metadata
Use the existing deployment module for deployment telemetry helpers instead of a standalone telemetry module.

Normalize the Sentry release value with the same deployment service version used on spans and logs so blank releases fall back to the Vercel commit SHA consistently.
