+++ packages/junior/src/deployment.ts
@@ -4,3 +4,30 @@ export const JUNIOR_CONVERSATION_WORK_CALLBACK_ROUTE =
   "/api/internal/agent/continue";
 export const LEGACY_JUNIOR_CONVERSATION_WORK_FUNCTION =
   "api/internal/agent/continue.ts";
+
+function toOptionalTrimmed(value: string | undefined): string | undefined {
+  const trimmed = value?.trim();
+  return trimmed ? trimmed : undefined;
+}
+
+/** Resolve the deployment version used for release and telemetry correlation. */
+export function getDeploymentServiceVersion(): string | undefined {
+  return (
+    toOptionalTrimmed(process.env.SENTRY_RELEASE) ??
+    toOptionalTrimmed(process.env.VERCEL_GIT_COMMIT_SHA)
+  );
+}
+
+/** Resolve deployment-scoped telemetry attributes from host environment. */
+export function getDeploymentTelemetryAttributes(): Record<string, string> {
+  const attributes: Record<string, string> = {};
+  const serviceVersion = getDeploymentServiceVersion();
+  const deploymentId = toOptionalTrimmed(process.env.VERCEL_DEPLOYMENT_ID);
+  if (serviceVersion) {
+    attributes["service.version"] = serviceVersion;
+  }
+  if (deploymentId) {
+    attributes["deployment.id"] = deploymentId;
+  }
+  return attributes;
+}
