+++ packages/junior/src/instrumentation.ts
@@ -1,5 +1,8 @@
 import * as Sentry from "@/chat/sentry";
-import { getDeploymentTelemetryAttributes } from "@/deployment-telemetry";
+import {
+  getDeploymentServiceVersion,
+  getDeploymentTelemetryAttributes,
+} from "@/deployment";
 
 function getSampleRate(value: string | undefined, fallback: number): number {
   if (!value) return fallback;
@@ -23,6 +26,7 @@ export function initSentry(): void {
 
   const dsn = process.env.SENTRY_DSN;
   const enableLogs = getBoolean(process.env.SENTRY_ENABLE_LOGS, Boolean(dsn));
+  const serviceVersion = getDeploymentServiceVersion();
   const deploymentSpanAttributes = getDeploymentTelemetryAttributes();
 
   Sentry.init({
@@ -31,7 +35,7 @@ export function initSentry(): void {
       process.env.SENTRY_ENVIRONMENT ??
       process.env.VERCEL_ENV ??
       process.env.NODE_ENV,
-    release: process.env.SENTRY_RELEASE ?? process.env.VERCEL_GIT_COMMIT_SHA,
+    release: serviceVersion,
     tracesSampleRate: getSampleRate(process.env.SENTRY_TRACES_SAMPLE_RATE, 1),
     sendDefaultPii: true,
     enabled: Boolean(dsn),
